requirements.txt
boto3==1.13.5
botocore==1.16.5
certifi==2020.4.5.1
chardet==3.0.4
Django==2.1.1
django-braces==1.14.0
django-crispy-forms==1.9.1
django-extensions==2.2.9
django-widget-tweaks==1.4.8
django-service-objects==0.7.0
django-bootstrap-modal-forms==1.5.0
docutils==0.15.2
ecdsa==0.15
envs==1.3
future==0.18.2
idna==2.9
jmespath==0.9.5
mock==4.0.2
pycryptodome==3.3.1
python-dateutil==2.8.1
python-jose-cryptodome==1.3.2
pytz==2020.1
Pillow==6.2.1
qrcode==6.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
warrant==0.6.1
django-bootstrap-modal-forms==1.5.0
dynamodb-json==1.3
simplejson==3.17.0
